Conflicts:
	campusipr_console/src/main/java/biz/mercue/campusipr/controller/TestController.java
	campusipr_console/src/main/resources/leo.yml
	campusipr_console/src/main/resources/server.yml

+++ campusipr_console/src/main/java/biz/mercue/campusipr/service/PatentServiceImpl.java
@@ -2,15 +2,24 @@ package biz.mercue.campusipr.service;
 
 
 
+import java.awt.print.Pageable;
 import java.util.List;
 
 import org.apache.log4j.Logger;
+import org.apache.poi.ss.formula.functions.Count;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import biz.mercue.campusipr.dao.PatentDao;
+import biz.mercue.campusipr.model.Business;
+import biz.mercue.campusipr.model.Country;
+import biz.mercue.campusipr.model.ListQueryForm;
 import biz.mercue.campusipr.model.Patent;
+import biz.mercue.campusipr.model.PatentContext;
+import biz.mercue.campusipr.model.PatentExtension;
+import biz.mercue.campusipr.model.PatentFamily;
+import biz.mercue.campusipr.util.Constants;
 import biz.mercue.campusipr.util.KeyGeneratorUtils;
 import biz.mercue.campusipr.util.StringUtils;
 
@@ -29,18 +38,42 @@ public class PatentServiceImpl implements PatentService{
 	public Patent getById(String businessId,String id) {
 		log.info("get by id: " + id);
 		Patent patent = patentDao.getById(businessId, id);
-		patent.getListBusiness().size();
-		//TODO change eger
-		patent.getListIPC().size();
-		patent.getListAgent().size();
-		patent.getListApplicant().size();
-		patent.getListAssignee().size();
-		patent.getListContact().size();
-		patent.getListCost().size();
-		patent.getListInventor().size();
-		patent.getListPortfolio().size();
+		if(patent!= null) {
+			PatentFamily family = patent.getFamily();
+			if(family!=null) {
+				log.info("family id :"+family.getPatent_family_id());
+			}else {
+				log.info("family is null");
+			}
+			PatentContext context = patent.getPatentContext();
+			if(context!=null) {
+				log.info("context id :"+context.getPatent_context_id());
+			}else {
+				log.info("context is null");
+			}
+			
+			PatentExtension extension = patent.getPatent_extension();
+			if(extension!=null) {
+				log.info("extension id :"+extension.getExtension_id());
+			}else {
+				log.info("extension is null");
+			}
+			
+			
+			
+			patent.getListBusiness().size();
+			//TODO change eger
+			patent.getListIPC().size();
+			patent.getListAgent().size();
+			patent.getListApplicant().size();
+			patent.getListAssignee().size();
+			patent.getListContact().size();
+			patent.getListCost().size();
+			patent.getListInventor().size();
+			patent.getListPortfolio().size();
+			
 
-		
+		}
 		return patent;
 	}
 	
@@ -54,13 +87,44 @@ public class PatentServiceImpl implements PatentService{
 	}
 
 	@Override
-	public void addPatent(Patent patent) {
+	public int addPatent(Patent patent) {
+		int taskResult= -1;
 		if(StringUtils.isNULL(patent.getPatent_id())) {
 			patent.setPatent_id(KeyGeneratorUtils.generateRandomString());
 		}
+		
+		if(patent.getBusiness() == null) {
+			return Constants.INT_DATA_ERROR;
+		}
+		
+		String applNo =  patent.getPatent_appl_no();
+		
+		if(!StringUtils.isNULL(applNo)) {
+			Patent appNoPatent = patentDao.getByApplNo(applNo);
+			if(appNoPatent!=null) {
+				List<Business> listBusiness = appNoPatent.getListBusiness();
+				for(Business business : listBusiness) {
+					if(patent.getBusiness().getBusiness_id().equals(business.getBusiness_id())) {
+						return Constants.INT_DATA_DUPLICATE;
+					}
+				}
+			}
+		}
+		
 
 		patentDao.create(patent);
+		return Constants.INT_SUCCESS;
 	}
+	
+	
+	@Override
+	public int addPatentByApplNo(Patent patent) {
+		
+		
+		return Constants.INT_CANNOT_FIND_DATA;
+	}
+	
+
 
 
 	@Override
@@ -112,37 +176,74 @@ public class PatentServiceImpl implements PatentService{
 
 			
 
-
+			return Constants.INT_SUCCESS;
+		}else {
+			return Constants.INT_CANNOT_FIND_DATA;
 		}
-		return 0;
 	}
 	
 
 	@Override
-	public void deletePatent(Patent patent) {
-		patentDao.delete(patent.getPatent_id());
+	public int deletePatent(Patent patent) {
+		Patent dbBean = patentDao.getById(patent.getPatent_id());
+		if(dbBean!=null) {
+			patentDao.delete(patent.getPatent_id());
+			return Constants.INT_SUCCESS;
+		}else {
+			return Constants.INT_CANNOT_FIND_DATA;
+		}
+		
 	}
 
-//	@Override
-//	public 	List<Patent> getByAdminId(String adminId,int page,int pageSize){
-//		return patentDao.getByAdminId(adminId,page,pageSize);
-//	}
 	
 	@Override
-	public 	List<Patent> getByBusinessId(String businessId,int page,int pageSize){
-		return patentDao.getByBusinessId(businessId,page,pageSize);
+	public 	ListQueryForm getByBusinessId(String businessId,int page){
+		
+		List list = patentDao.getByBusinessId(businessId,page,Constants.SYSTEM_PAGE_SIZE);
+		int count = patentDao.getCountByBusinessId(businessId);
+		ListQueryForm form = new ListQueryForm(count,Constants.SYSTEM_PAGE_SIZE,list);
+		
+		return form;
 	}
 	
 	
 	@Override
-	public List<Patent> getByPatentIds(List<String> list){
-		return patentDao.getByPatentIds(list);
+	public List<Patent> getByPatentIds(List<String> idList,String businessId){
+		List list = patentDao.getByPatentIds(idList,businessId);
+
+		return list;
 	}
 	
 	@Override
 	public Patent getByPatentNo(String patentNo){
 		return patentDao.getByPatentNo(patentNo);
 	}
+
+
+	@Override
+	public ListQueryForm searchPatent(String text, String businessId, int page) {
+		
+		//TODO no finish yet
+		List list = patentDao.getByBusinessId(businessId,page,Constants.SYSTEM_PAGE_SIZE);
+		int count = patentDao.getCountByBusinessId(businessId);
+		ListQueryForm form = new ListQueryForm(count,Constants.SYSTEM_PAGE_SIZE,list);
+		
+		return form;
+	}
+	
+	@Override
+	public ListQueryForm fieldSearchPatent(String text, String businessId, int page) {
+		
+		//TODO no finish yet
+		List list = patentDao.getByBusinessId(businessId,page,Constants.SYSTEM_PAGE_SIZE);
+		int count = patentDao.getCountByBusinessId(businessId);
+		ListQueryForm form = new ListQueryForm(count,Constants.SYSTEM_PAGE_SIZE,list);
+		
+		return form;
+	}
+	
+	
+
 
 	
 }
